requirements.txt
pandas
openpyxl
xlsxwriter
streamlit
plotly
altair
schedule
pytz
supabase
python-dotenv